Break up helpers + add a few (potentially)
